requirements.txt
discord.py
fabric
